requirements.txt
# ── IA / Gemini ──────────────────────────────────────────
google-generativeai>=0.8.0       # SDK oficial de Gemini API

# ── WordPress Integration ────────────────────────────────
requests>=2.31.0                 # Llamadas a WordPress REST API

# ── Interfaz Gráfica ─────────────────────────────────────
streamlit>=1.33.0                # GUI web local

# ── Scraping de Amazon (para parsear URLs) ───────────────
beautifulsoup4>=4.12.0           # Parsear HTML de páginas de producto
httpx>=0.27.0                    # Cliente HTTP (scraping de Amazon)
dnspython>=2.6.0                 # Resolver DNS alternativo (bypass bloqueo DNS corporativo)

# ── Utilidades ───────────────────────────────────────────
python-dotenv>=1.0.0             # Cargar variables de entorno desde .env
pydantic>=2.0.0                  # Validación de datos y modelos
tenacity>=8.2.0                  # Retry automático para llamadas a API
loguru>=0.7.0                    # Logging estructurado
